+++ includes/reader-activation/class-integrations.php
@@ -547,3 +547,4 @@ class Integrations {
 		}
 	}
 }
+Integrations::init();