+++ src/wizards/audience/views/content-gates/types/index.d.ts
@@ -143,9 +143,14 @@ type MeteringCountdownConfig = {
 	cta_product_id: number;
 };
 
+type AdvancedSettingsConfig = {
+	restrict_feeds: boolean;
+};
+
 type GateSettings = {
 	content_gifting?: ContentGiftingConfig;
 	countdown_banner?: MeteringCountdownConfig;
+	advanced_settings?: AdvancedSettingsConfig;
 };
 
 type GateConfig = {
@@ -166,6 +171,6 @@ type Institution = {
 		np_institution_reader_data: string;
 	};
 	_embedded?: {
-		'wp:featuredmedia'?: Array< { source_url: string } >;
+		'wp:featuredmedia'?: Array<{ source_url: string }>;
 	};
 };